Fix: 오류 수정

+++ server/photomory/src/main/java/com/example/photomory/repository/MyAlbumRepository.java
@@ -5,8 +5,8 @@ import org.springframework.data.jpa.repository.JpaRepository;
 
 import java.util.List;
 
-public interface MyAlbumRepository extends JpaRepository<MyAlbum, Long> {
+public interface MyAlbumRepository extends JpaRepository<MyAlbum, Integer> {
 
-    List<MyAlbum> findByUserId(Long userId);
+    List<MyAlbum> findByUserId(long userId);
 
 }
